Write about multilinear polynomial evaluation
The new material starts by introducing a general form for bases of the
space of k-variate multilinear polynomials over a field. This general
form includes as special cases both the monomial basis and the Lagrange
bases, so algorithms can be written to cover both of these cases and
more.

We introduce and analyze two evaluation algorithms which allow streaming
of the polynomial coefficients in arbitrary order, and discuss a method
of interpolating between them to make a flexible space-time tradeoff. We
next present an evaluation algorithm with better performance, but which
requires ordered or random access to the polynomial coefficients.